Remove old GenShell
